Completed Chapter 5 Representing Code
In this chapter I created an Abstract Syntax Tree to represent the structure of passed text. Right now we created a brute force main function that prints a binary expression: () * (-123) (group 45.67)). While not correct solution from what is expected in the book. At least the code runs a correct test.